requirements.txt
asgiref==3.7.2
astroid==2.15.5
certifi==2023.5.7
charset-normalizer==3.2.0
dill==0.3.6
Django==4.2.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==10.0.0
platformdirs==3.5.1
pylint==2.17.4
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.3
urllib3==2.0.4
wrapt==1.15.0
